ruby: PulseAudio: implement DRC
This is a very crude implementation. In my machine, when requesting 20ms
latency PA gives me a 15ms buffer and 5ms minimum request.
I avoid trying to deduce the hardware ring buffer is 5ms long. But it
may actually be a good idea, I just couldn't bother to test it in
detail.
In any case this means level(), when requesting 20ms latency in my
machine, returns:
- 100%: When it's 66% - 100% full
- 66%: When it's 33% - 66% full
- 33%: When it's 0% - 33% full
- 0%: When it's 0% full, and even the hardware buffer is getting
  depleted

Notice the 66% case is specially bad since it tells to DRC it needs to
send less frames when it actually may need more (it may be 33%).

Decreasing the PA minimum request could improve all this. But in
practice it seems to work well enough in my machine with 60ms+ latency,
so I'm leaving any improvements to somebody else.